+++ docs/en/changes/changes.md
@@ -55,6 +55,7 @@
 * Fix table exists check in the JDBC Storage Plugin.
 * Fix day-based table rolling time range strategy in JDBC storage.
 * Add `maxInboundMessageSize (SW_DCS_MAX_INBOUND_MESSAGE_SIZE)` configuration to change the max inbound message size of DCS.
+* Fix Service Layer when building Events in the EventHookCallback.
 
 #### UI